requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.1.1
kafka-python==2.0.2
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
joblib==1.4.2
fastapi==0.115.5
uvicorn==0.32.1
openai==1.55.3
